ver13
Zmienione sprzedawanie energii i produkowanie energii z wartosci float na wartosci int

+++ BuildMenu.h
@@ -5,12 +5,11 @@
 #include <string>
 #include <optional>
 
-// Struktura pojedynczego elementu do zbudowania
 struct BuildItem {
     int                  id;
     std::string          name;
-    float                basePrice;
-    float                currentPrice;
+    int                  basePrice;
+    int                  currentPrice;
     sf::RectangleShape   buttonShape;
     sf::Text             nameText;
     sf::Text             priceText;
@@ -21,19 +20,16 @@ public:
     BuildMenu();
     ~BuildMenu() = default;
 
-    // Struktura przechowująca informacje o przeciąganym obiekcie
     struct DragState {
         int itemId;
-        float price;
+        int price;
         sf::Sprite sprite;
     };
 
-    // Inicjalizacja przyjmuje teraz referencję do tekstur wczytanych w main
     bool initialize(const sf::Font& font, const std::vector<sf::Texture>& textures);
-
     void setVisible(bool visible);
     bool isVisible() const;
-    void handleEvent(const sf::Event& ev, sf::RenderWindow& window, float& money);
+    void handleEvent(const sf::Event& ev, sf::RenderWindow& window, int& money);
     void draw(sf::RenderWindow& window);
 
     std::optional<DragState>& getDragState();
@@ -43,10 +39,7 @@ private:
     std::vector<BuildItem> items;
     sf::RectangleShape     background;
     bool                   visible;
-
-    // Wskaźnik do tekstur, których właścicielem jest main.cpp
     const std::vector<sf::Texture>* itemTexturesRef = nullptr;
-
     std::optional<DragState> dragState;
 
     void setupItems(const sf::Font& font);
