+++ HUD.h
@@ -8,18 +8,15 @@ class HUD {
 public:
     HUD(sf::Font& font);
 
-    void update(float money, float energy, float maxEnergy, float envHealth, float totalTime);
+    // ZMIANA: Metoda przyjmuje teraz wartości int
+    void update(int money, int energy, int maxEnergy, float envHealth, float totalTime);
     void draw(sf::RenderWindow& window);
-
-    // ZMIANA: Metoda do sprawdzania, czy przycisk został kliknięty
     const sf::FloatRect getEnergySellButtonBounds() const;
 
 private:
     sf::Font& m_font;
     sf::Text m_labelMoney, m_labelEnergy, m_labelEnv, m_labelTime;
     sf::Text m_valueMoney, m_valueEnergy, m_valueEnv;
-
-    // ZMIANA: Elementy przycisku sprzedaży energii
     sf::CircleShape m_energySellButtonBg;
     sf::Text m_energySellIcon;
 };
